Fix gitleaks warnings without fragile exclusions
The rest of the codebase uses Amazon's default credential when examples are needed.
Gitleaks appears to recognise this and doesn't warn on those credentials.
If we fix the gitleaks warning by using gitleaks ignore, which appears to reference specific
lines/parts of code, then it seems like this is liable to break if that code gets moved around.
Also .gitleaksignore is labeled as experimental by the documentation.