Add superlint on PR and on merge to main (#2)
* Add superlint on PR and merge to main

* Trim down the super-linter config

In an ideal world I think it makes sense to have most of these enabled but right now they're
not things we've been linting on before, so disabling them lets us make forward progress.

* Fix gitleaks warnings without fragile exclusions

The rest of the codebase uses Amazon's default credential when examples are needed.
Gitleaks appears to recognise this and doesn't warn on those credentials.
If we fix the gitleaks warning by using gitleaks ignore, which appears to reference specific
lines/parts of code, then it seems like this is liable to break if that code gets moved around.
Also .gitleaksignore is labeled as experimental by the documentation.

---------

Co-authored-by: cchin118 <cchin118@bloomberg.net>