Replace usage of lua-penlight with our own functions
Lua-penlight in its entirety is upwards of 13k lines of code and of that we use a total of 2 functions.
Managing dependencies (in particular for lua as its script and needs to be available at an appropriate path
at runtime) is a pain. In this case I'm not sure that pain is worth not having to write these two very
simple functions ourselves and having to manage a dependency instead.