Improve haproxy deactivate robustness (#8)
* Check if all added files exist before deactivating

This helps prevent us from getting into a half-deactivated state where added files have been removed along with
the base commit, but where patches have not yet been extracted.

* Force-remove patches to avoid errors

If we manage to get into a half-deactivated state, we might not have any patches. This would already
require manual intervention to fix, but adding the force just reduces the effort slightly of getting
out of that situation when it arises, and there is little reason not to.

* Update the haproxy readme with upgrade instructions

---------

Co-authored-by: Jacques Heunis <jheunis@bloomberg.net>