Improve (de)activation usability (#11)
* Add extra comments to the (de)activate scripts

This is just to add further justification for each step along the way.

* Copy extra files before applying patches during activation

As explained in the comment, we do this because copying is much less likely to fail than patch application.
When patch application fails its probably because you're changing the base commit, in which case it would
be helpful if, once you'd manually applied each problematic patch, the haproxy source directory was in a
consistent, fully-activated state. This would mean that you could then simply deactivate again and get
back a full set of updated patches that you can commit or apply again by activating.

Without this, if patches fail to apply you'll need to manually copy all the added files.