Add a fix for the freq_ctr overshoot calculation overflowing with high limits (#12)
Bandwidth limits in haproxy are almost always stored or passed around as uints internally. The one exception to that is
freq_ctr_overshoot_period, which calculates how far over the limit a freq_ctr has gone (producing the minimum amount of
time that the user should wait before submitting more events). This result is returned currently as a signed integer.
This is not inherently problematic, but it's a little strange (all other quantities are unsigned in those calculations,
and they're all closely related) and also the conversion to a signed quantity happens implicitly right at the end,
after a possible overflow of the unsigned intermediate value.

The end result of this for us is that if you set a very high limit when using the freq_ctr-based bandwidth limiting,
then in some cases you will be instructed to wait for a long time, causing transfers to stall.